Fixed not being able to delete warps. Fixed error with wars not saving Fixed pitch and yaw not being loaded. Updated to 3.6

+++ src/main/java/com/collinsrichard/easywarp/managers/FileManager.java
@@ -8,7 +8,10 @@ import org.bukkit.Location;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.ObjectInputStream;
 import java.util.ArrayList;
 
 public class FileManager {
@@ -41,8 +44,6 @@ public class FileManager {
                     for (String i : parse) {
                         try {
                             WarpManager.addWarp(WarpManager.deserialize(i));
-
-                            ois.close();
                         } catch (Exception e) {
                             System.out.println("Easy Warp had an error loading warps.");
                         }
@@ -54,8 +55,30 @@ public class FileManager {
         }
     }
 
+    public void saveWarps() {
+        for(String key : warpConfig.getKeys(true)){
+            warpConfig.set(key, null);
+        }
+
+        for (Warp warp : WarpManager.getWarps()) {
+            warpConfig.set("warps." + warp.getName() + ".world", warp.getLocation().getWorld().getName());
+            warpConfig.set("warps." + warp.getName() + ".x", warp.getLocation().getX());
+            warpConfig.set("warps." + warp.getName() + ".y", warp.getLocation().getY());
+            warpConfig.set("warps." + warp.getName() + ".z", warp.getLocation().getZ());
+
+            warpConfig.set("warps." + warp.getName() + ".yaw", warp.getLocation().getYaw());
+            warpConfig.set("warps." + warp.getName() + ".pitch", warp.getLocation().getPitch());
+        }
+
+        try {
+            warpConfig.save(warpConfigFile);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
     public void loadWarps() {
-        if (getOldFile().exists()) {
+        if (getOldFile().exists() && !warpConfigFile.exists()) {
             loadWarpsOld();
             getOldFile().delete();
             saveWarps();
@@ -72,38 +95,18 @@ public class FileManager {
 
             Location warpLocation = new Location(Bukkit.getWorld(worldName), x, y, z);
 
-            if (warpConfig.contains("yaw")) {
+            if (warpConfig.contains("warps." + name + ".yaw")) {
                 Float yaw = Float.parseFloat(warpConfig.getString("warps." + name + ".yaw"));
                 warpLocation.setYaw(yaw);
             }
 
-            if (warpConfig.contains("pitch")) {
+            if (warpConfig.contains("warps." + name + ".pitch")) {
                 Float pitch = Float.parseFloat(warpConfig.getString("warps." + name + ".pitch"));
                 warpLocation.setYaw(pitch);
             }
 
             Warp warp = new Warp(name, warpLocation);
             WarpManager.addWarp(warp);
-        }
-    }
-
-    public void saveWarps() {
-        warpConfigFile.delete();
-
-        for (Warp warp : WarpManager.getWarps()) {
-            warpConfig.set("warps." + warp.getName() + ".world", warp.getLocation().getWorld().getName());
-            warpConfig.set("warps." + warp.getName() + ".x", warp.getLocation().getX());
-            warpConfig.set("warps." + warp.getName() + ".y", warp.getLocation().getY());
-            warpConfig.set("warps." + warp.getName() + ".z", warp.getLocation().getZ());
-
-            warpConfig.set("warps." + warp.getName() + ".yaw", warp.getLocation().getYaw());
-            warpConfig.set("warps." + warp.getName() + ".pitch", warp.getLocation().getPitch());
-        }
-
-        try {
-            warpConfig.save(warpConfigFile);
-        } catch (IOException e) {
-            e.printStackTrace();
         }
     }
 }
